Drowsyness-Detection/Drowsyness_Detection_TranferLearning_Inception/app/requirements.txt
numpy==1.26.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.0.0
pygame==2.6.1
streamlit==1.41.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
